MBX-3168: AppDelegate LifeCycle real example
